backend/scrapper/requirements.txt
# Web scraping dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Supabase client
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
